Every int turned to byte

+++ 18GhostsGame/Player.cs
@@ -11,12 +11,12 @@ namespace _18GhostsGame
          Blue Ghosts = ghosts[2,~];
          Yellow Ghosts = ghosts[1,~];
         */
-        private int[,] ghosts;
+        private byte[,] ghosts;
         private string playerNum;
 
         public Player(byte playerNum)
         {
-            ghosts = new int[3, 3] { { 2, 0, 0 }, { 0, 6, 0 }, { 0, 12, 0 } };
+            ghosts = new byte[3, 3] { { 2, 0, 0 }, { 0, 6, 0 }, { 0, 12, 0 } };
             this.playerNum = playerNum == 1 ? "Player 1" : "Player 2";
         }
 
@@ -32,32 +32,32 @@ namespace _18GhostsGame
 
         public void SetGhostsToZero()
         {
-            ghosts = new int[3, 3] { { 0, 21, 0 }, { 0, 0, 8 }, { 0, 0, 0 } };
+            ghosts = new byte[3, 3] { { 0, 21, 0 }, { 0, 0, 8 }, { 0, 0, 0 } };
         }
 
         public string GetPlayerNum() => playerNum;
 
-        public int[,] GetGhosts() => ghosts;
-        public int[] GetGhosts(string wichGhosts)
+        public byte[,] GetGhosts() => ghosts;
+        public byte[] GetGhosts(string wichGhosts)
         {
-            int[] toReturn = new int[3];
-            int[,] toReturnAll = new int[3, 3];
+            byte[] toReturn = new byte[3];
+            byte[,] toReturnAll = new byte[3, 3];
 
             switch (wichGhosts)
             {
                 case "red":
                     toReturn =
-                        new int[3] { ghosts[0, 0], ghosts[0, 1], ghosts[0, 2] };
+                        new byte[3] { ghosts[0, 0], ghosts[0, 1], ghosts[0, 2] };
                     break;
 
                 case "blue":
                     toReturn =
-                        new int[3] { ghosts[1, 0], ghosts[1, 1], ghosts[1, 2] };
+                        new byte[3] { ghosts[1, 0], ghosts[1, 1], ghosts[1, 2] };
                     break;
 
                 case "yellow":
                     toReturn =
-                        new int[3] { ghosts[2, 0], ghosts[2, 1], ghosts[2, 2] };
+                        new byte[3] { ghosts[2, 0], ghosts[2, 1], ghosts[2, 2] };
                     break;
 
                 case "all":
